Nested asset's composed charts are preserved

+++ app/components/data-menu/services/asset-service.js
@@ -19,7 +19,9 @@ angular.module('data-menu')
 
           if (timeseriesInAsset || selection.asset === asset.entity_name + "$" + asset.id) {
             // Remove
-            ChartCompositionService.removeSelection(selection.uuid);
+            if (State.context === 'dashboard') {
+              ChartCompositionService.removeSelection(selection.uuid);
+            }
           } else {
             // Keep
             keepSelections.push(selection);
